adding more extensive test in AdvectBMI_cpp_test.cpp
