+++ src/VarManager.h
@@ -79,19 +79,6 @@ public:
 	void Gfw_Var();
 	void GridCellCount_Var();
 	void NthSelectedOutput_Var();
-	void OutputVarsAddSolutionProperties_Var();
-	void OutputVarsAddSolutionTotalMolalities_Var();
-	void OutputVarsAddSolutionMolalities_Var();
-	void OutputVarsAddSolutionActivities_Var();
-	void OutputVarsAddExchangeMolalities_Var();
-	void OutputVarsAddSurfaceMolalities_Var();
-	void OutputVarsAddEquilibriumPhases_Var();
-	void OutputVarsAddSaturationIndices_Var();
-	void OutputVarsAddGases_Var();
-	void OutputVarsAddKineticReactants_Var();
-	void OutputVarsAddSolidSolutions_Var();
-	void OutputVarsAddCalculateValues_Var();
-	void OutputVarsFinalize_Var();
 	void Saturation_Var();
 	void SelectedOutput_Var();
 	void SelectedOutputColumnCount_Var();
